+++ frontend/src/pages/Login.tsx
@@ -1,5 +1,3 @@
-import { useEffect } from "react";
-
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
@@ -15,7 +13,7 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { loginApi, useAuthStatus, useRoutes } from "@/hooks/use-api";
+import { loginApi, useAuthStatus } from "@/hooks/use-api";
 import { useAuthStore } from "@/store/use-auth";
 import { tryCatch } from "@/utils/try-catch";
 
@@ -27,74 +25,6 @@ const formSchema = z.object({
     message: "Password must be at least 2 characters.",
   }),
 });
-
-const nowSchema = z.object({
-  title: z.string().min(2, {
-    message: "Title least 2 characters.",
-  }),
-  desc: z.string().min(2, {
-    message: "Description must be at least 2 characters.",
-  }),
-});
-
-export function AddNow() {
-  // const { setIsAuth } = useAuthStore();
-  const { data } = useRoutes();
-  const form = useForm<z.infer<typeof nowSchema>>({
-    resolver: zodResolver(nowSchema),
-    defaultValues: {
-      title: "",
-      desc: "",
-    },
-  });
-
-  async function onSubmit(values: z.infer<typeof nowSchema>) {
-    const body = {
-      title: values.title,
-      desc: values.desc,
-    };
-
-    await data.now.post.mutateAsync({
-      body,
-    });
-
-    form.reset();
-  }
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-        <FormField
-          control={form.control}
-          name="title"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Title</FormLabel>
-              <FormControl>
-                <Input placeholder="title..." {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="desc"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Description</FormLabel>
-              <FormControl>
-                <Input placeholder="description..." {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <Button type="submit">Submit</Button>
-      </form>
-    </Form>
-  );
-}
 
 export function LoginForm() {
   const { setIsAuth } = useAuthStore();
@@ -194,10 +124,5 @@ export const Login = () => {
   if (error) console.warn(error);
   if (data) setIsAuth(true);
 
-  return (
-    <div>
-      {isAuth ? <LogOutButton /> : <LoginForm />}
-      {isAuth && <AddNow />}
-    </div>
-  );
+  return <div>{isAuth ? <LogOutButton /> : <LoginForm />}</div>;
 };
